apps/langchain-hfgti/requirements.txt
langchain
text_generation
sentence-transformers==2.2.2
kafka-python